% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augbin_latent_function.R
\name{augbinrheum}
\alias{augbinrheum}
\title{Augmented binary method}
\usage{
augbinrheum(data = NULL, cts = 1, bin = 1, dichot = NULL)
}
\arguments{
\item{data}{A dataset to analyse. Must be an object of class data.frame
details. Defaults to NULL.}

\item{cts}{Number of continuous components within the composite endpoint. Allows for
1 or 2 continuous components. Defaults to 1.}

\item{bin}{Number of binary components within the composite endpoint. Allows for
0 or 1 binary components. Any additional components can be collapsed and accounted for within
this indicator. Defaults to 1.}

\item{dichot}{Vector of dichotomisation threshold(s) for continuous component(s).
For more than one continuous components values must be entered as e.g. c(0,0). 
Defaults to NULL.}
}
\description{
Fits the augmented binary method using an underlying latent
variable model allowing for up to two continuous components 
and one binary component, common in rheumatic conditions.
}
\details{
The responder outcome is assumed to define a responder when
continuous components are below a particular dichotomisation threshold and
a binary component is 0.

Available as a Shiny app with documentation at \href{https://github.com/martinamcm/AugBin}{https://github.com/martinamcm/AugBin}
}
\examples{
Example data available at \href{https://github.com/martinamcm/AugBin}{https://github.com/martinamcm/AugBin} for cts=1 and bin=1
odds_ratio <- augbin()$odds_ratio
risk_ratio <- augbin()$risk_ratio
risk_diff <- augbin()$risk_diff
response_prob <- augbin()$response_prob
}
\references{
McMenamin M, Grayling MJ, Berglind A, Wason JMS. Increasing power in the 
analysis of responder endpoints in rheumatology: a software tutorial.
medRxiv. 2020. \href{https://doi.org/10.1101/2020.07.28.20163378}{https://doi.org/10.1101/2020.07.28.20163378}
}
\author{
Martina McMenamin.
}
