% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\title{Graphical user interface to multi-arm trial design determination}
\usage{
gui()
}
\description{
\code{gui()} run an R Shiny web browser based graphical user interface for
augbinrheum. For details of app functionality, see \href{https://github.com/martinamcm/AugBin}{AugBin}
}
\examples{
# Launch the graphical user interface
\dontrun{gui()}
}
