% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\title{Graphical user interface to analyse composite endpoints using the augmented binary method}
\usage{
gui()
}
\description{
\code{gui()} runs an R Shiny web browser based graphical user interface for
\code{\link{augbinrheum}}. For details of Shiny app functionality, 
see \href{https://github.com/martinamcm/AugBin}{AugBin}
}
\examples{
# Launch the graphical user interface
\dontrun{gui()}
}
